+++ src/server/electron-window.js
@@ -79,7 +79,7 @@ module.exports = function(options={}) {
 
     window.webContents.once('page-title-updated', ()=>{
         if (options.fullscreen) {
-            window.webContents.sendInputEvent({keyCode: 'F11', type: 'keyDown'})
+            window.webContents.executeJavaScript('window.ELECTRON_FULLSCREEN()', true)
         }
         if (app._noGpu) {
             window.webContents.executeJavaScript(`